refactor: move out and set it as customhook
