+++ src/hooks/useMediaQuery.ts
@@ -1,0 +1,20 @@
+import { useEffect, useState } from "react";
+
+export function useMediaQuery(query: string) {
+	const [matches, setMatches] = useState(() => {
+		if (typeof window !== "undefined") {
+			return window.matchMedia(query).matches;
+		}
+		return false;
+	});
+
+	useEffect(() => {
+		const media = window.matchMedia(query);
+		const listener = (event: MediaQueryListEvent) => setMatches(event.matches);
+
+		media.addEventListener("change", listener);
+		return () => media.removeEventListener("change", listener);
+	}, [query]);
+
+	return matches;
+}
